refactor: extract CLI program builder into src/cli/program.ts
